% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grupOzetC}
\alias{grupOzetC}
\title{C++ ile Hızlı Grup Ortalaması Hesaplama}
\usage{
grupOzetC(x, grup)
}
\arguments{
\item{x}{Değerlerin bulunduğu sayısal vektör.}

\item{grup}{Grupları belirten karakter vektörü.}
}
\value{
Grup isimlerini ve ortalamalarını içeren bir liste.
}
\description{
Bir vektörün değerlerini belirtilen gruplara göre gruplayarak
ortalamalarını hızlıca hesaplar.
}
