% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hareketliOrtalamaC}
\alias{hareketliOrtalamaC}
\title{C++ ile Hızlı Hareketli Ortalama (Rolling Mean) Hesaplama}
\usage{
hareketliOrtalamaC(x, n)
}
\arguments{
\item{x}{Sayısal vektör.}

\item{n}{Pencere boyutu (pozitif bir tam sayı).}
}
\value{
Giriş vektörüyle aynı uzunlukta hareketli ortalamaları içeren vektör. İlk n-1 eleman NA olacaktır.
}
\description{
Bu fonksiyon, sayısal bir vektör üzerinde belirtilen pencere boyutuna göre
hızlıca hareketli ortalama hesaplar.
}
\examples{
veri <- c(1, 2, 4, 8, 16, 32)
hareketliOrtalamaC(veri, n = 3)
}
