% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{paralelUzaklikC}
\alias{paralelUzaklikC}
\title{C++ ve OpenMP ile Paralel Mesafe Matrisi Hesaplama}
\usage{
paralelUzaklikC(X, threads = 0L)
}
\arguments{
\item{X}{Sayısal bir matris (satırlar gözlemleri, sütunlar özellikleri temsil eder).}

\item{threads}{Kullanılacak iş parçacığı sayısı. 0 verilirse sistemdeki tüm çekirdekler kullanılır.}
}
\value{
N x N boyutunda bir mesafe matrisi.
}
\description{
Bu fonksiyon, verilen bir matrisin satırları arasındaki
Öklid mesafelerini paralel olarak hesaplar.
}
\examples{
matris <- matrix(runif(100), nrow = 10)
paralelUzaklikC(matris, threads = 2)
}
