% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vektorKareC}
\alias{vektorKareC}
\title{C++ ile Vektörün Karesini Alan Hızlı Fonksiyon}
\usage{
vektorKareC(x)
}
\arguments{
\item{x}{Sayısal vektör}
}
\description{
C++ ile Vektörün Karesini Alan Hızlı Fonksiyon
}
\examples{
vektorKareC(c(1, 2, 3, 4))
}
